requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
Unipath
gunicorn
whitenoise
dj_database_url
django-widget-tweaks
